requirements.txt
numpy>=1.26
scipy>=1.11
pandas>=2.0
librosa>=0.10
soundfile>=0.12
scikit-learn>=1.3
joblib>=1.3
matplotlib>=3.8
seaborn>=0.13
python-dotenv>=1.0
